[dagster-airlift] show airflow-side retries test behavior

+++ python_modules/libraries/dagster-airlift/dagster_airlift/test/shared_fixtures.py
@@ -8,10 +8,12 @@ from pathlib import Path
 from tempfile import TemporaryDirectory
 from typing import Any, Callable, Generator, List, Optional
 
+import psutil
 import pytest
 import requests
 from dagster._core.test_utils import environ
 from dagster._time import get_current_timestamp
+from dagster._utils import process_is_alive
 
 from dagster_airlift.core.airflow_instance import AirflowInstance
 from dagster_airlift.core.basic_auth import AirflowBasicAuthBackend
@@ -100,8 +102,9 @@ def stand_up_airflow(
         assert airflow_ready, "Airflow did not start within 30 seconds..."
         yield process
     finally:
-        # Kill process group, since process.kill and process.terminate do not work.
-        os.killpg(process.pid, signal.SIGKILL)
+        if process_is_alive(process.pid):
+            # Kill process group, since process.kill and process.terminate do not work.
+            os.killpg(process.pid, signal.SIGKILL)
 
 
 @pytest.fixture(name="airflow_instance")
@@ -179,7 +182,9 @@ def stand_up_dagster(
         assert dagster_ready, "Dagster did not start within 30 seconds..."
         yield process
     finally:
-        os.killpg(process.pid, signal.SIGKILL)
+        if psutil.Process(pid=process.pid).is_running():
+            # Kill process group, since process.kill and process.terminate do not work.
+            os.killpg(process.pid, signal.SIGKILL)
 
 
 ####################################################################################################
